Implemented main menu, viewmodels and scene loading!
- Still buggy with animation loading...

+++ SynapseEngine/Engine/Registry/Pool/Storage/Mixin/Flag/Core/AtomicFlagMixin.h
@@ -26,7 +26,7 @@ namespace Syn
 
     public:
         SYN_INLINE void PushFlagImpl() {
-            constexpr uint8_t mask = (1 << REGENERATE_BIT) | (1 << UPDATE_BIT) | (1 << INDEX_CHANGED_BIT);
+            constexpr uint32_t mask = (1 << REGENERATE_BIT) | (1 << UPDATE_BIT) | (1 << INDEX_CHANGED_BIT);
             _flags.emplace_back(mask);
             _state.fetch_or(mask, std::memory_order_relaxed);
         }
